+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `toListOfType` extension for JsonArray
 * `toObjectOfType` extension for JsonObject
 * Directory (`klib.files.Directory`)
+* `readAll` extension for FileList
+* `writeAll` extensions for FileList
 ### Changed
 * Moved and Renamed a lot of packages (Major Breaking Change)
 * `toObjectFromType` extension on String uses Generics